Cleanups: Code Cleanup AstType.METHOD_TYPE -> AstType.METHOD Removed AstClassRef and unified with AstType.REF

+++ jtransc-main/test/javatest/lang/BasicTypesTest.java
@@ -20,8 +20,8 @@ public class BasicTypesTest {
 	private static void swapTests() {
 		System.out.println(Short.reverseBytes((short) 0x1234));
 		System.out.println(Short.reverseBytes((short) 0xF5F2));
-		System.out.println(Integer.reverseBytes((int) 0x12345678));
-		System.out.println(Integer.reverseBytes((int) 0xF2345678));
+		System.out.println(Integer.reverseBytes(0x12345678));
+		System.out.println(Integer.reverseBytes(0xF2345678));
 		System.out.println(0x123456789ABCDEF0L);
 		System.out.println(-1234567890);
 		System.out.println(Long.reverseBytes(0x123456789ABCDEF0L));
