+++ jtransc-rt-core/src/jtransc/JTranscBits.java
@@ -278,7 +278,7 @@ public class JTranscBits {
 	}
 
 	static public int readInt32BE(byte[] data) {
-		return (int)((data[0] << 24) + (data[1] << 16) + (data[2] << 8) + (data[3] << 0));
+		return (data[0] << 24) + (data[1] << 16) + (data[2] << 8) + (data[3] << 0);
 	}
 
 	public static int clamp(int v, int min, int max) {
